imported-content/adventure/legacy/requirements.txt
cmd2
openai
pyyaml
pillow
flask
flask_cors
gunicorn
gevent

# Development dependencies
pytest
pytest-cov
